# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Task Management API
  description: API for managing authentication, comments, tasks, and task assignees
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Local development server
paths:
...
  /comments:
    post:
      summary: Add a new comment
      description: Adds a new comment to a specific task.
      tags:
        - Comments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequestDto'
      responses:
        "201":
          description: Comment added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponseDto'
  /comments/{id}:
    get:
      summary: Get comment by ID
      description: Fetches a comment by its unique ID.
      tags:
        - Comments
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identifier of the comment
          example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        "200":
          description: Comment fetched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponseDto'

    delete:
      summary: Delete a comment
      description: Deletes a comment by its unique ID.
      tags:
        - Comments
      parameters:
        - name: id
          in: path
          required: true
          description: Unique identifier of the comment
          example: "123e4567-e89b-12d3-a456-426614174002"
      responses:
        "204":
          description: Comment deleted successfully
  /comments/task/{taskId}:
    get:
      summary: Get comments by task ID
      description: Fetches all comments associated with a specific task.
      tags:
        - Comments
      parameters:
        - name: taskId
          in: path
          required: true
          description: Unique identifier of the task
          example: "123e4567-e89b-12d3-a456-426614174001"
        - name: pageable
          in: query
          required: false
          description: Pagination and sorting information
          schema:
            type: string
          example: "{ \"page\": 0, \"size\": 10, \"sort\": \"task,desc\" }"
      responses:
        "200":
          description: Fetched comments for the task
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommentResponseDto'
...
components:
  schemas:
...
    CommentResponseDto:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        task:
          $ref: '#/components/schemas/TaskResponseDto'
        author:
          $ref: '#/components/schemas/UserResponseDto'
    TaskResponseDto:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        author:
          $ref: '#/components/schemas/UserResponseDto'
        taskAssignees:
          type: array
          items:
            $ref: '#/components/schemas/TaskAssigneeResponseDto'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        comments:
          type: array
          items:
            $ref: '#/components/schemas/CommentResponseDto'
    CreateCommentRequestDto:
      type: object
      properties:
        content:
          type: string
        taskId:
          type: string
...
    TaskAssigneeResponseDto:
      type: object
      properties:
        id:
          type: string
        task:
          $ref: '#/components/schemas/TaskResponseDto'
        assignee:
          $ref: '#/components/schemas/UserResponseDto'
        assignedAt:
          type: string
          format: date-time
...
    UserResponseDto:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string